Web: Files: refactoring: created Badges component, refactoring EditingWrapperComponent, FilesTileContent, FilesRowContent
