+++ products/ASC.Files/Client/src/components/pages/Home/Section/Body/sub-components/icons.js
@@ -1,10 +1,7 @@
-import React from "react";
 import styled from "styled-components";
 
 import commonIconsStyles from "@appserver/components/utils/common-icons-style";
 
-import CheckIcon from "../../../../../../../public/images/check.react.svg";
-import CrossIcon from "../../../../../../../../../../public/images/cross.react.svg";
 import FavoriteIcon from "../../../../../../../public/images/favorite.react.svg";
 import FileActionsConvertEditDocIcon from "../../../../../../../public/images/file.actions.convert.edit.doc.react.svg";
 import FileActionsLockedIcon from "../../../../../../../public/images/file.actions.locked.react.svg";
@@ -17,30 +14,6 @@ export const EncryptedFileIcon = styled.div`
   margin-top: 14px;
   margin-left: ${(props) => (props.isEdit ? "40px" : "12px")};
 `;
-
-const StyledCheckIcon = styled(CheckIcon)`
-  ${commonIconsStyles}
-  path {
-    fill: #a3a9ae;
-  }
-  :hover {
-    fill: #657077;
-  }
-`;
-export const okIcon = <StyledCheckIcon className="edit-ok-icon" size="scale" />;
-
-const StyledCrossIcon = styled(CrossIcon)`
-  ${commonIconsStyles}
-  path {
-    fill: #a3a9ae;
-  }
-  :hover {
-    fill: #657077;
-  }
-`;
-export const cancelIcon = (
-  <StyledCrossIcon className="edit-cancel-icon" size="scale" />
-);
 
 export const StyledFavoriteIcon = styled(FavoriteIcon)`
   ${commonIconsStyles}
